+++ examples/_example_utils.py
@@ -8,13 +8,13 @@ from rich.text import Text
 from combustion_ble.ble_data.probe_temperatures import ProbeTemperatures
 from combustion_ble.devices.device import Device
 from combustion_ble.devices.meat_net_node import MeatNetNode
-from combustion_ble.devices.probe import Probe
+from combustion_ble.devices.probe import Probe, VirtualTemperatures
 from combustion_ble.logger import LOGGER
 
 
 def configure_logging(log_level="DEBUG", bleak_log_level="WARNING"):
     FORMAT = "%(message)s"
-    logging.basicConfig(level="NOTSET", format=FORMAT, datefmt="[%X]", handlers=[RichHandler()])
+    logging.basicConfig(level="ERROR", format=FORMAT, datefmt="[%X]", handlers=[RichHandler()])
     bleak_logger = logging.getLogger("bleak")
     bleak_logger.setLevel(bleak_log_level)
     LOGGER.setLevel(log_level)
@@ -33,6 +33,10 @@ def format_connection_state(connection_state: str = "unknown"):
 
 def format_temperatures(temperature_data: ProbeTemperatures | None):
     return f"Temps: {str([round(t, 1) for t in (temperature_data.values if temperature_data else [])])}"
+
+
+def format_virtual_temperatures(temperature: VirtualTemperatures):
+    return f"Core: {round(temperature.core_temperature, 1)} | Ambient: {round(temperature.ambient_temperature, 1)} | Surface: {round(temperature.surface_temperature, 1)}"
 
 
 def format_device_name(device: Device):