+++ examples/read_prediction.py
@@ -6,7 +6,7 @@ from rich.live import Live
 from rich.table import Table
 
 from combustion_ble.device_manager import DeviceManager
-from examples.example_utils import configure_logging, format_device_name
+from examples._example_utils import configure_logging, format_device_name
 
 
 async def main():
@@ -29,7 +29,7 @@ async def main():
             prediction = "N/A"
             if probe.prediction_info:
                 state = probe.prediction_info.prediction_state.to_string()
-                prediction = str(probe.prediction_info)
+                prediction = str(probe._prediction_info)
             table.add_row(name, str(probe.last_update_time), state, prediction)
 
         return table